+++ src/main/java/com/picimako/terra/wdio/screenshot/gutter/TerraScreenshotValidationLineMarkerProvider.java
@@ -47,10 +47,10 @@ import com.picimako.terra.wdio.screenshot.TerraScreenshotCollector;
  * <p>
  * In case the call references a non-existent screenshot, the line marker is not added.
  * Instead they will be marked by the {@link com.picimako.terra.wdio.screenshot.inspection.MissingScreenshotInspection}.
+ *
+ * @since 0.4.0
  */
 public class TerraScreenshotValidationLineMarkerProvider extends RelatedItemLineMarkerProvider {
-
-    private final TerraScreenshotCollector collector = new TerraScreenshotCollector();
 
     @Override
     protected void collectNavigationMarkers(@NotNull PsiElement element, @NotNull Collection<? super RelatedItemLineMarkerInfo<?>> result) {
@@ -60,7 +60,7 @@ public class TerraScreenshotValidationLineMarkerProvider extends RelatedItemLine
             if (nameArgument == null) {
                 JSExpression methodExpression = terraCallExpr.getMethodExpression();
                 if (methodExpression != null) {
-                    PsiElement[] screenshots = collector.collectForDefault(methodExpression);
+                    PsiElement[] screenshots = new TerraScreenshotCollector(element.getProject()).collectForDefault(methodExpression);
                     if (screenshots.length > 0) {
                         PsiElement leafElement = findLeafElement(terraCallExpr);
                         if (leafElement != null) {