+++ src/test/java/com/picimako/terra/wdio/screenshot/TerraScreenshotNameResolverTest.java
@@ -35,67 +35,75 @@ public class TerraScreenshotNameResolverTest extends BasePlatformTestCase {
     //resolveName
 
     public void testResolveName() {
-        myFixture.configureByFile("tests/wdio/nameResolution/resolveName-spec.js");
-        JSLiteralExpression element = (JSLiteralExpression) myFixture.getFile().findElementAt(myFixture.getCaretOffset()).getParent();
-
+        JSLiteralExpression element = configureFileForJSLiteralExpression("tests/wdio/nameResolution/resolveName-spec.js");
         assertThat(new TerraScreenshotNameResolver().resolveName(element)).isEqualTo("terra_screenshot[with_name].png");
     }
 
     public void testResolveNameToEmptyStringIfNoParentDescribeCall() {
-        myFixture.configureByFile("tests/wdio/nameResolution/resolveNameNoParentDescribe-spec.js");
-
-        JSLiteralExpression element = (JSLiteralExpression) myFixture.getFile().findElementAt(myFixture.getCaretOffset()).getParent();
-
+        JSLiteralExpression element = configureFileForJSLiteralExpression("tests/wdio/nameResolution/resolveNameNoParentDescribe-spec.js");
         assertThat(new TerraScreenshotNameResolver().resolveName(element)).isEmpty();
     }
 
     public void testResolveNameToEmptyStringIfParentDescribeCallHasNoNameParameter() {
-        myFixture.configureByFile("tests/wdio/nameResolution/resolveNameNoParentDescribeName-spec.js");
-
-        JSLiteralExpression element = (JSLiteralExpression) myFixture.getFile().findElementAt(myFixture.getCaretOffset()).getParent();
-
+        JSLiteralExpression element = configureFileForJSLiteralExpression("tests/wdio/nameResolution/resolveNameNoParentDescribeName-spec.js");
         assertThat(new TerraScreenshotNameResolver().resolveName(element)).isEmpty();
+    }
+
+    //resolveName with partial test id
+
+    public void testResolvePartialNameWithTestId() {
+        JSLiteralExpression element = configureFileForJSLiteralExpression("tests/wdio/nameResolution/resolveNameWithTestId-spec.js");
+        assertThat(new TerraScreenshotNameResolver().resolveName(element)).isEqualTo("terra_screenshot[test_id].png");
+    }
+
+    public void testResolvePartialNameWithMultipleTestIds() {
+        JSLiteralExpression element = configureFileForJSLiteralExpression("tests/wdio/nameResolution/resolveNameWithMultipleTestIds-spec.js");
+        assertThat(new TerraScreenshotNameResolver().resolveName(element)).isEqualTo("terra_screenshot[test_id]_and_another_[testid].png");
+    }
+
+    public void testResolveFullNameWithNonMatchingPartialTestId() {
+        JSLiteralExpression element = configureFileForJSLiteralExpression("tests/wdio/nameResolution/resolveNameWithNonMatchingTestId-spec.js");
+        assertThat(new TerraScreenshotNameResolver().resolveName(element)).isEqualTo("terra_screenshot[with_name_[tes)t_id]].png");
     }
 
     //resolveDefaultName
 
     public void testResolveDefaultName() {
-        myFixture.configureByFile("tests/wdio/nameResolution/resolveDefaultName-spec.js");
-        JSExpression element = (JSExpression) myFixture.getFile().findElementAt(myFixture.getCaretOffset()).getParent();
-
+        JSExpression element = configureFileForJSExpression("tests/wdio/nameResolution/resolveDefaultName-spec.js");
         assertThat(new TerraScreenshotNameResolver().resolveDefaultName(element)).isEqualTo("terra_screenshot[default].png");
     }
 
     public void testResolveDefaultNameToEmptyStringIfNoParentDescribeCall() {
-        myFixture.configureByFile("tests/wdio/nameResolution/resolveDefaultNameNoParentDescribe-spec.js");
-
-        JSExpression element = (JSExpression) myFixture.getFile().findElementAt(myFixture.getCaretOffset()).getParent();
-
+        JSExpression element = configureFileForJSExpression("tests/wdio/nameResolution/resolveDefaultNameNoParentDescribe-spec.js");
         assertThat(new TerraScreenshotNameResolver().resolveDefaultName(element)).isEqualTo("");
     }
 
     public void testResolveDefaultNameToEmptyStringIfParentDescribeCallHasNoNameParameter() {
-        myFixture.configureByFile("tests/wdio/nameResolution/resolveDefaultNameNoParentDescribeName-spec.js");
-
-        JSExpression element = (JSExpression) myFixture.getFile().findElementAt(myFixture.getCaretOffset()).getParent();
-
+        JSExpression element = configureFileForJSExpression("tests/wdio/nameResolution/resolveDefaultNameNoParentDescribeName-spec.js");
         assertThat(new TerraScreenshotNameResolver().resolveDefaultName(element)).isEqualTo("");
     }
 
     //resolveWithFallback
 
     public void testResolveByLiteralNoFallback() {
-        myFixture.configureByFile("tests/wdio/nameResolution/resolveName-spec.js");
-        JSLiteralExpression element = (JSLiteralExpression) myFixture.getFile().findElementAt(myFixture.getCaretOffset()).getParent();
-
+        JSLiteralExpression element = configureFileForJSLiteralExpression("tests/wdio/nameResolution/resolveName-spec.js");
         assertThat(new TerraScreenshotNameResolver().resolveWithFallback(element, null)).isEqualTo("terra_screenshot[with_name].png");
     }
 
     public void testResolveByMethodExpressionFallback() {
-        myFixture.configureByFile("tests/wdio/nameResolution/resolveDefaultName-spec.js");
-
-        JSExpression element = (JSExpression) myFixture.getFile().findElementAt(myFixture.getCaretOffset()).getParent();
-
+        JSExpression element = configureFileForJSExpression("tests/wdio/nameResolution/resolveDefaultName-spec.js");
         assertThat(new TerraScreenshotNameResolver().resolveWithFallback(null, element)).isEqualTo("terra_screenshot[default].png");
+    }
+
+    //Helper methods
+
+    private JSExpression configureFileForJSExpression(String path) {
+        myFixture.configureByFile(path);
+        return (JSExpression) myFixture.getFile().findElementAt(myFixture.getCaretOffset()).getParent();
+    }
+
+    private JSLiteralExpression configureFileForJSLiteralExpression(String path) {
+        myFixture.configureByFile(path);
+        return (JSLiteralExpression) myFixture.getFile().findElementAt(myFixture.getCaretOffset()).getParent();
     }
 }