+++ src/test/java/com/picimako/terra/wdio/screenshot/reference/TerraScreenshotReferenceTest.java
@@ -16,6 +16,9 @@
 
 package com.picimako.terra.wdio.screenshot.reference;
 
+import static com.picimako.terra.wdio.ScreenshotTypeHelper.diff;
+import static com.picimako.terra.wdio.ScreenshotTypeHelper.latest;
+import static com.picimako.terra.wdio.ScreenshotTypeHelper.reference;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.Set;
@@ -25,17 +28,11 @@ import com.intellij.psi.PsiPolyVariantReference;
 import com.intellij.psi.ResolveResult;
 import com.intellij.testFramework.fixtures.BasePlatformTestCase;
 
-import com.picimako.terra.wdio.TerraWdioFolders;
-
 /**
  * Unit test for {@link TerraScreenshotReference}.
  */
 public class TerraScreenshotReferenceTest extends BasePlatformTestCase {
 
-    private static final Set<String> RESOLVED_FILE_PATHS = Set.of(
-        "/src/tests/wdio/__snapshots__/reference/en/chrome_medium/some-spec/terra-_screenshot--[with-_-replaced-_-characters_-].png",
-        "/src/tests/wdio/__snapshots__/reference/en/chrome_huge/some-spec/terra-_screenshot--[with-_-replaced-_-characters_-].png"
-    );
     private static final Set<String> LOCATION_STRINGS = Set.of("en/chrome/huge", "en/chrome/medium");
 
     @Override
@@ -43,102 +40,138 @@ public class TerraScreenshotReferenceTest extends BasePlatformTestCase {
         return "testdata/terra/projectroot";
     }
 
+    //Terra.describeViewports
+
     public void testReferencesForTerraDescribeViewportsAndValidatesScreenshot() {
-        validateReferencesForSourceFile("ScreenshotResolveTerraDescribeViewportsValidatesScreenshot-spec.js");
+        validateReferencesForSourceFile("ScreenshotResolveTerraDescribeViewportsValidatesScreenshot");
     }
 
     public void testReferencesForTerraDescribeViewportsAndValidatesElement() {
-        validateReferencesForSourceFile("ScreenshotResolveTerraDescribeViewportsValidatesElement-spec.js");
+        validateReferencesForSourceFile("ScreenshotResolveTerraDescribeViewportsValidatesElement");
     }
 
     public void testReferencesForTerraDescribeViewportsAndItMatchesScreenshot() {
-        validateReferencesForSourceFile("ScreenshotResolveTerraDescribeViewportsItMatchesScreenshot-spec.js");
+        validateReferencesForSourceFile("ScreenshotResolveTerraDescribeViewportsItMatchesScreenshot");
     }
 
     public void testReferencesForTerraDescribeViewportsAndItValidatesElement() {
-        validateReferencesForSourceFile("ScreenshotResolveTerraDescribeViewportsItValidatesElement-spec.js");
+        validateReferencesForSourceFile("ScreenshotResolveTerraDescribeViewportsItValidatesElement");
     }
 
+    //describe
+
     public void testReferencesForDescribeAndValidatesScreenshot() {
-        validateReferencesForSourceFile("ScreenshotResolveDescribeValidatesScreenshot-spec.js");
+        validateReferencesForSourceFile("ScreenshotResolveDescribeValidatesScreenshot");
     }
 
     public void testReferencesForDescribeAndValidatesElement() {
-        validateReferencesForSourceFile("ScreenshotResolveDescribeValidatesElement-spec.js");
+        validateReferencesForSourceFile("ScreenshotResolveDescribeValidatesElement");
     }
 
     public void testReferencesForDescribeAndItMatchesScreenshot() {
-        validateReferencesForSourceFile("ScreenshotResolveDescribeItMatchesScreenshot-spec.js");
+        validateReferencesForSourceFile("ScreenshotResolveDescribeItMatchesScreenshot");
     }
 
     public void testReferencesForDescribeAndItValidatesElement() {
-        validateReferencesForSourceFile("ScreenshotResolveDescribeItValidatesElement-spec.js");
+        validateReferencesForSourceFile("ScreenshotResolveDescribeItValidatesElement");
     }
 
+    //Terra.describeViewports + describe
+
     public void testReferencesForTerraDescribeViewportsDescribeAndValidatesScreenshot() {
-        validateReferencesForSourceFile("ScreenshotResolveDescribeViewportsDescribeValidatesScreenshot-spec.js");
+        validateReferencesForSourceFile("ScreenshotResolveDescribeViewportsDescribeValidatesScreenshot");
     }
 
     public void testReferencesForTerraDescribeViewportsDescribeAndValidatesElement() {
-        validateReferencesForSourceFile("ScreenshotResolveDescribeViewportsDescribeValidatesElement-spec.js");
+        validateReferencesForSourceFile("ScreenshotResolveDescribeViewportsDescribeValidatesElement");
     }
 
     public void testReferencesForTerraDescribeViewportsDescribeAndItMatchesScreenshot() {
-        validateReferencesForSourceFile("ScreenshotResolveDescribeViewportsDescribeItMatchesScreenshot-spec.js");
+        validateReferencesForSourceFile("ScreenshotResolveDescribeViewportsDescribeItMatchesScreenshot");
     }
 
     public void testReferencesForTerraDescribeViewportsDescribeAndItValidatesElement() {
-        validateReferencesForSourceFile("ScreenshotResolveDescribeViewportsDescribeItValidatesElement-spec.js");
+        validateReferencesForSourceFile("ScreenshotResolveDescribeViewportsDescribeItValidatesElement");
     }
 
+    //describe + describe
+
     public void testReferencesForDescribeDescribeAndValidatesScreenshot() {
-        validateReferencesForSourceFile("ScreenshotResolveDescribeDescribeValidatesScreenshot-spec.js");
+        validateReferencesForSourceFile("ScreenshotResolveDescribeDescribeValidatesScreenshot");
     }
 
     public void testReferencesForDescribeDescribeAndValidatesElement() {
-        validateReferencesForSourceFile("ScreenshotResolveDescribeDescribeValidatesElement-spec.js");
+        validateReferencesForSourceFile("ScreenshotResolveDescribeDescribeValidatesElement");
     }
 
     public void testReferencesForDescribeDescribeAndItMatchesScreenshot() {
-        validateReferencesForSourceFile("ScreenshotResolveDescribeDescribeItMatchesScreenshot-spec.js");
+        validateReferencesForSourceFile("ScreenshotResolveDescribeDescribeItMatchesScreenshot");
     }
 
     public void testReferencesForDescribeDescribeAndItValidatesElement() {
-        validateReferencesForSourceFile("ScreenshotResolveDescribeDescribeItValidatesElement-spec.js");
+        validateReferencesForSourceFile("ScreenshotResolveDescribeDescribeItValidatesElement");
     }
 
-    public void testSingleReference() {
-        configureSpecFile("ScreenshotResolveSingleResult-spec.js");
-        configureSingleReferenceScreenshot();
+    //test id
+
+    public void testScreenshotResolveDescribeValidatesElementTestId() {
+        myFixture.configureByFile("tests/wdio/ScreenshotResolveDescribeValidatesElementTestId-spec.js");
+        myFixture.copyFileToProject(reference("/en/chrome_huge/ScreenshotResolveDescribeValidatesElementTestId-spec/terra_screenshot[single].png"));
 
         PsiElement element = myFixture.getFile().findElementAt(myFixture.getCaretOffset()).getParent();
         TerraScreenshotPsiFile terraScreenshot = (TerraScreenshotPsiFile) element.getReferences()[0].resolve();
 
-        assertThat(terraScreenshot.getVirtualFile().getPath()).isEqualTo("/src/tests/wdio/__snapshots__/reference/en/chrome_huge/some-spec/terra_screenshot[single].png");
+        assertThat(terraScreenshot.getVirtualFile().getPath())
+            .isEqualTo("/src/tests/wdio/__snapshots__/reference/en/chrome_huge/ScreenshotResolveDescribeValidatesElementTestId-spec/terra_screenshot[single].png");
+    }
+
+    public void testScreenshotResolveDescribeValidatesElementNoTestId() {
+        myFixture.configureByFile("tests/wdio/ScreenshotResolveDescribeValidatesElementNoTestId-spec.js");
+        myFixture.copyFileToProject(reference("/en/chrome_huge/ScreenshotResolveDescribeValidatesElementNoTestId-spec/terra_screenshot[i_am_[not_)_single]].png"));
+
+        PsiElement element = myFixture.getFile().findElementAt(myFixture.getCaretOffset()).getParent();
+        TerraScreenshotPsiFile terraScreenshot = (TerraScreenshotPsiFile) element.getReferences()[0].resolve();
+
+        assertThat(terraScreenshot.getVirtualFile().getPath())
+            .isEqualTo("/src/tests/wdio/__snapshots__/reference/en/chrome_huge/ScreenshotResolveDescribeValidatesElementNoTestId-spec/terra_screenshot[i_am_[not_)_single]].png");
+    }
+
+    //single result
+
+    public void testSingleReference() {
+        myFixture.configureByFile("tests/wdio/ScreenshotResolveSingleResult-spec.js");
+        myFixture.copyFileToProject(reference("/en/chrome_huge/ScreenshotResolveSingleResult-spec/terra_screenshot[single].png"));
+
+        PsiElement element = myFixture.getFile().findElementAt(myFixture.getCaretOffset()).getParent();
+        TerraScreenshotPsiFile terraScreenshot = (TerraScreenshotPsiFile) element.getReferences()[0].resolve();
+
+        assertThat(terraScreenshot.getVirtualFile().getPath())
+            .isEqualTo("/src/tests/wdio/__snapshots__/reference/en/chrome_huge/ScreenshotResolveSingleResult-spec/terra_screenshot[single].png");
         assertThat(terraScreenshot.getPresentation().getLocationString()).isEqualTo("en/chrome/huge");
     }
 
-    private void validateReferencesForSourceFile(String fileName) {
-        configureSpecFile(fileName);
-        configureMultipleReferenceScreenshots();
+    //Helper methods
+
+    private void validateReferencesForSourceFile(String specName) {
+        myFixture.configureByFile("tests/wdio/" + specName + "-spec.js");
+        myFixture.copyFileToProject(reference("/en/chrome_huge/" + specName + "-spec/terra-_screenshot--[with-_-replaced-_-characters_-].png"));
+        myFixture.copyFileToProject(reference("/en/chrome_medium/" + specName + "-spec/terra-_screenshot--[with-_-replaced-_-characters_-].png"));
+        myFixture.copyFileToProject(diff("/en/chrome_huge/" + specName + "-spec/terra-_screenshot--[with-_-replaced-_-characters_-].png"));
+        myFixture.copyFileToProject(latest("/en/chrome_huge/" + specName + "-spec/terra-_screenshot--[with-_-replaced-_-characters_-].png"));
 
         PsiElement element = myFixture.getFile().findElementAt(myFixture.getCaretOffset()).getParent();
         ResolveResult[] resolveResults = ((PsiPolyVariantReference) element.getReferences()[0]).multiResolve(false);
 
         assertThat(resolveResults).hasSize(2);
-        assertThat(RESOLVED_FILE_PATHS).containsExactlyInAnyOrder(getPath(resolveResults[0]), getPath(resolveResults[1]));
+        assertThat(resolvedFilePaths(specName)).containsExactlyInAnyOrder(getPath(resolveResults[0]), getPath(resolveResults[1]));
         assertThat(LOCATION_STRINGS).containsExactlyInAnyOrder(getLocationString(resolveResults[0]), getLocationString(resolveResults[1]));
     }
 
-    private void configureSpecFile(String fileName) {
-        myFixture.configureByFile("tests/wdio/" + fileName);
-    }
-
-    private void configureMultipleReferenceScreenshots() {
-        myFixture.copyFileToProject("tests/wdio/__snapshots__/reference/en/chrome_huge/some-spec/terra-_screenshot--[with-_-replaced-_-characters_-].png");
-        myFixture.copyFileToProject("tests/wdio/__snapshots__/reference/en/chrome_medium/some-spec/terra-_screenshot--[with-_-replaced-_-characters_-].png");
-        myFixture.copyFileToProject("tests/wdio/__snapshots__/diff/en/chrome_huge/some-spec/terra-_screenshot--[with-_-replaced-_-characters_-].png");
-        myFixture.copyFileToProject("tests/wdio/__snapshots__/latest/en/chrome_huge/some-spec/terra-_screenshot--[with-_-replaced-_-characters_-].png");
+    private Set<String> resolvedFilePaths(String spec) {
+        return Set.of(
+            "/src/tests/wdio/__snapshots__/reference/en/chrome_medium/" + spec + "-spec/terra-_screenshot--[with-_-replaced-_-characters_-].png",
+            "/src/tests/wdio/__snapshots__/reference/en/chrome_huge/" + spec + "-spec/terra-_screenshot--[with-_-replaced-_-characters_-].png"
+        );
     }
 
     private String getPath(ResolveResult resolveResult) {
@@ -147,9 +180,5 @@ public class TerraScreenshotReferenceTest extends BasePlatformTestCase {
 
     private String getLocationString(ResolveResult resolveResult) {
         return ((TerraScreenshotPsiFile) resolveResult.getElement()).getPresentation().getLocationString();
-    }
-
-    private void configureSingleReferenceScreenshot() {
-        myFixture.copyFileToProject("tests/wdio/__snapshots__/reference/en/chrome_huge/some-spec/terra_screenshot[single].png");
     }
 }